Move idependent components to the components folder
