+++ app/public/components/flags/flagVisualization.js
